IMPROVE: plugin code layout
